ref: fold the two transient sets onto a shared base
`TransientHashSet` and `TransientSortedSet` were the largest clone in
`src/php`. A previous pass left them alone because merging would have baked in
the missing `__invoke` on the sorted side; now that the asymmetry is gone the
two are identical apart from how they print and which persistent set they
freeze into, so both extend `HashSet\AbstractTransientSet` and supply just
those two methods. No flag or mode argument is involved, and it mirrors
`SortedMap\PersistentSortedMap` extending `Map\AbstractPersistentMap`.
