+++ src/php/Lang/Collections/HashSet/TransientHashSet.php
@@ -4,78 +4,16 @@ declare(strict_types=1);
 
 namespace Phel\Lang\Collections\HashSet;
 
-use Phel\Lang\Collections\Map\TransientMapInterface;
-use Phel\Lang\HasherInterface;
-use Stringable;
-
 /**
  * @template V
  *
- * @implements TransientHashSetInterface<V>
+ * @extends AbstractTransientSet<V>
  */
-final readonly class TransientHashSet implements TransientHashSetInterface, Stringable
+final readonly class TransientHashSet extends AbstractTransientSet
 {
-    /**
-     * @param TransientMapInterface<V, V> $transientMap
-     */
-    public function __construct(
-        private HasherInterface $hasher,
-        private TransientMapInterface $transientMap,
-    ) {}
-
     public function __toString(): string
     {
         return '<TransientSet count=' . $this->transientMap->count() . '>';
-    }
-
-    /**
-     * Membership lookup so transient sets remain callable like their persistent
-     * counterparts: `((transient #{:a}) :a) ; => :a`, else `nil`.
-     *
-     * @param V $key
-     *
-     * @return V|null
-     */
-    public function __invoke(mixed $key): mixed
-    {
-        return $this->transientMap->find($key);
-    }
-
-    public function count(): int
-    {
-        return $this->transientMap->count();
-    }
-
-    /**
-     * @param mixed $key
-     */
-    public function contains($key): bool
-    {
-        return $this->transientMap->contains($key);
-    }
-
-    /**
-     * @param V $value
-     *
-     * @return TransientHashSetInterface<V>
-     */
-    public function add($value): TransientHashSetInterface
-    {
-        $this->transientMap->put($value, $value);
-
-        return $this;
-    }
-
-    /**
-     * @param V $value
-     *
-     * @return TransientHashSetInterface<V>
-     */
-    public function remove($value): TransientHashSetInterface
-    {
-        $this->transientMap->remove($value);
-
-        return $this;
     }
 
     /**
